feat: return reviewable model upon upload
